Fetch grandes cantidades de información arreglado (ya se puede descargar ficheros grandes mediante API routes con múltiples llamadas)
